[qca-nss-dp] IPQ95xx: Support PTP service code timestamp
In RX path, the packet with the ptp service code is received,
this packet will be delivered to the phy driver api rxtstamp
for handing rx timestamping.

In TX path, the socket option SKBTX_HW_TSTAMP is enabled, the
packet will be delivered to the phy driver api txtstamp for
handing tx timestamping.

Change-Id: I51bb66942c0bb622fb91a3d65ad322b16c863020
Signed-off-by: Luo Jie <quic_luoj@quicinc.com>