nss_dp_main: Use a 'phy-handle' property to connect to the PHY
The original method of connecting a PHY to the ethernet controller
requires the "qcom,link-poll", and "qcom,phy-mdio-addr" devicetree
properties. This is redundant. The PHY node already contains the MDIO
address, and attaching a PHY implies "link-poll".

Allow using a "phy-handle" property. Remove the following properties,
as they are no longer used:
    * "qcom,link-poll"
    * "qcom,phy-mdio-addr"
    * "mdio-bus"
    * "qcom,forced-speed"
    * "qcom,forced-duplex"

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>
Signed-off-by: Robert Marko <robimarko@gmail.com>