[qca-nss-dp] Fix the packets processing in the Rx NAPI
Fix the packet processing through the list_for_each_entry_safe() API
in the Rx NAPI poll function.
The 2nd parameter of this API should be a temporary variable which is
required by this API for its internal processing. If this variable
(2nd parameter) needs to be used by the caller, then the 'head' check
should be done on it prior to its usage, otherwise we might face the
data abort issues.

Change-Id: Ic3f7728ac504fb6650055d10302d69a11a14442b
Signed-off-by: Manish Verma <quic_maniverm@quicinc.com>