[qca-nss-dp] Add support for the XGMAC latency computation
When the timestamp measurement Kernel config is enabled, compute and
store the latency of the ingress packet from the XGMAC interface.

This latency information can be useful for figuring out the packet's
latency from the point it enqueued in the XGMAC interface till it
becomes available in the Rx NAPI poll function for its processing.

Change-Id: Ie125e64586129b2e1383b825db45896cc609ccf3
Signed-off-by: Manish Verma <quic_maniverm@quicinc.com>