nss_dp_main: support fixed-link
Add support for "fixed link" to support ethernet MACs which are not
connected to an MDIO managed PHY. For example, IPQ5018 has two internal
MACs, one connected to a GE PHY and the other connected to a Uniphy to
support connections to an external switch over a fixed link. As such,
check for a fixed link in absence of a phy-handle and return an error
when no phy-handle and fixed link is found.

Signed-off-by: Ziyang Huang <hzyitc@outlook.com>
Signed-off-by: George Moussalem <george.moussalem@outlook.com>